refactor the main documentation site
This patch moves the fails of the main part of the documentation
into a new folder docs_main.
The contents of those pages are added.
New sections are created for rfsm and bfl.
The main menu on the left column when is reordered to have a
clearer structure.
A TODO file is added to track wip.